add build constraints to limit syscalls to linux
