+++ internal/util/syscall.go
@@ -1,3 +1,7 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+// +build linux
+
 package util
 
 import "syscall"
